refactor without test the action and observation space, need to test now
