+++ grid2op/Observation.py
@@ -44,11 +44,13 @@ import pdb
 try:
     from .Exceptions import *
     from .Reward import ConstantReward, RewardHelper
+    from .Space import SerializableSpace
     from ._utils import extract_from_dict, save_to_dict
 except (ModuleNotFoundError, ImportError):
     from Exceptions import *
     from Reward import ConstantReward, RewardHelper
     from _utils import extract_from_dict, save_to_dict
+    from Space import SerializableSpace
 
 # TODO be able to change reward here
 
@@ -1451,420 +1453,24 @@ class CompleteObservation(Observation):
         return res
 
 
-class SerializableObservationSpace:
-    """
-    This class allows to serialize / de serialize the observation space.
-
-    It should not be used inside an Environment, as some functions of the Observation might not be compatible with
-    the serialization, for example the "forecast" method.
-
-    Attributes
-    ----------
-    n_lines: :class:`int`
-        number of powerline in the _grid
-
-    n_gen: :class:`int`
-        number of generators in the _grid
-
-    n_load: :class:`int`
-        number of loads in the powergrid
-
-    subs_info: :class:`numpy.array`, dtype:int
-        for each substation, gives the number of elements connected to it
-
-    load_to_subid: :class:`numpy.array`, dtype:int
-        for each load, gives the id the substation to which it is connected
-
-    gen_to_subid: :class:`numpy.array`, dtype:int
-        for each generator, gives the id the substation to which it is connected
-
-    lines_or_to_subid: :class:`numpy.array`, dtype:int
-        for each lines, gives the id the substation to which its "origin" end is connected
-
-    lines_ex_to_subid: :class:`numpy.array`, dtype:int
-        for each lines, gives the id the substation to which its "extremity" end is connected
-
-    load_to_sub_pos: :class:`numpy.array`, dtype:int
-        The topology if of the subsation *i* is given by a vector, say *sub_topo_vect* of size
-        :attr:`Backend.subs_info`\[i\]. For a given load of id *l*, :attr:`Backend._load_to_sub_pos`\[l\] is the index
-        of the load *l* in the vector *sub_topo_vect*. This means that, if
-        *sub_topo_vect\[ action._load_to_sub_pos\[l\] \]=2*
-        then load of id *l* is connected to the second bus of the substation.
-
-    gen_to_sub_pos: :class:`numpy.array`, dtype:int
-        same as :attr:`Backend._load_to_sub_pos` but for generators.
-
-    lines_or_to_sub_pos: :class:`numpy.array`, dtype:int
-        same as :attr:`Backend._load_to_sub_pos`  but for "origin" end of powerlines.
-
-    lines_ex_to_sub_pos: :class:`numpy.array`, dtype:int
-        same as :attr:`Backend._load_to_sub_pos` but for "extremity" end of powerlines.
-
-    load_pos_topo_vect: :class:`numpy.array`, dtype:int
-        It has a similar role as :attr:`Backend._load_to_sub_pos` but it gives the position in the vector representing
-        the whole topology. More concretely, if the complete topology of the powergrid is represented here by a vector
-        *full_topo_vect* resulting of the concatenation of the topology vector for each substation
-        (see :attr:`Backend._load_to_sub_pos`for more information). For a load of id *l* in the powergrid,
-        :attr:`Backend._load_pos_topo_vect`\[l\] gives the index, in this *full_topo_vect* that concerns load *l*.
-        More formally, if *_topo_vect\[ backend._load_pos_topo_vect\[l\] \]=2* then load of id l is connected to the
-        second bus of the substation.
-
-    gen_pos_topo_vect: :class:`numpy.array`, dtype:int
-        same as :attr:`Backend._load_pos_topo_vect` but for generators.
-
-    lines_or_pos_topo_vect: :class:`numpy.array`, dtype:int
-        same as :attr:`Backend._load_pos_topo_vect` but for "origin" end of powerlines.
-
-    lines_ex_pos_topo_vect: :class:`numpy.array`, dtype:int
-        same as :attr:`Backend._load_pos_topo_vect` but for "extremity" end of powerlines.
-
-    observationClass: ``type``
-        Class used to build the observations. It defaults to :class:`CompleteObservation`
-
-    """
-    def __init__(self,
-                 n_gen, n_load, n_lines, subs_info,
-                 load_to_subid, gen_to_subid, lines_or_to_subid, lines_ex_to_subid,
-                 load_to_sub_pos, gen_to_sub_pos, lines_or_to_sub_pos, lines_ex_to_sub_pos,
-                 load_pos_topo_vect, gen_pos_topo_vect, lines_or_pos_topo_vect, lines_ex_pos_topo_vect,
-                 observationClass=CompleteObservation):
-
-        # print("ObservationHelper init with reward_helper class: {}".format(self.reward_helper.template_reward))
-        self.n_gen = n_gen
-        self.n_load = n_load
-        self.n_lines = n_lines
-        self.subs_info = subs_info
-        self.dim_topo = np.sum(subs_info)
-
-        # to which substation is connected each element
-        self.load_to_subid = load_to_subid
-        self.gen_to_subid = gen_to_subid
-        self.lines_or_to_subid = lines_or_to_subid
-        self.lines_ex_to_subid = lines_ex_to_subid
-        # which index has this element in the substation vector
-        self.load_to_sub_pos = load_to_sub_pos
-        self.gen_to_sub_pos = gen_to_sub_pos
-        self.lines_or_to_sub_pos = lines_or_to_sub_pos
-        self.lines_ex_to_sub_pos = lines_ex_to_sub_pos
-        # which index has this element in the topology vector
-        self.load_pos_topo_vect = load_pos_topo_vect
-        self.gen_pos_topo_vect = gen_pos_topo_vect
-        self.lines_or_pos_topo_vect = lines_or_pos_topo_vect
-        self.lines_ex_pos_topo_vect = lines_ex_pos_topo_vect
-
-        self.observationClass = observationClass
-
-        self.empty_obs = self.observationClass(n_gen=self.n_gen, n_load=self.n_load, n_lines=self.n_lines,
-                                               subs_info=self.subs_info, dim_topo=self.dim_topo,
-                                               load_to_subid=self.load_to_subid,
-                                               gen_to_subid=self.gen_to_subid,
-                                               lines_or_to_subid=self.lines_or_to_subid,
-                                               lines_ex_to_subid=self.lines_ex_to_subid,
-                                               load_to_sub_pos=self.load_to_sub_pos,
-                                               gen_to_sub_pos=self.gen_to_sub_pos,
-                                               lines_or_to_sub_pos=self.lines_or_to_sub_pos,
-                                               lines_ex_to_sub_pos=self.lines_ex_to_sub_pos,
-                                               load_pos_topo_vect=self.load_pos_topo_vect,
-                                               gen_pos_topo_vect=self.gen_pos_topo_vect,
-                                               lines_or_pos_topo_vect=self.lines_or_pos_topo_vect,
-                                               lines_ex_pos_topo_vect=self.lines_ex_pos_topo_vect,
-                                               obs_env=None,
-                                               action_helper=None)
-
-        self.n = self.empty_obs.size()
-
-    @staticmethod
-    def from_dict(dict_):
-        """
-        Allows the de-serialization of an object stored as a dictionnary (for example in the case of json saving).
-
-        Parameters
-        ----------
-        dict_: ``dict``
-            Representation of an Observation Space (aka ObservartionHelper) as a dictionnary.
-
-        Returns
-        -------
-        res: :class:``SerializableObservationSpace``
-            An instance of an observationHelper matching the dictionnary.
-
-        """
-        if isinstance(dict_, str):
-            path = dict_
-            if not os.path.exists(path):
-                raise Grid2OpException("Unable to find the file \"{}\" to load the ObservationSpace".format(path))
-            with open(path, "r", encoding="utf-8") as f:
-                dict_ = json.load(fp=f)
-
-        n_gen = extract_from_dict(dict_, "n_gen", int)
-        n_load = extract_from_dict(dict_, "n_load", int)
-        n_lines = extract_from_dict(dict_, "n_lines", int)
-
-        subs_info = extract_from_dict(dict_, "subs_info", lambda x: np.array(x).astype(np.int))
-        load_to_subid = extract_from_dict(dict_, "load_to_subid", lambda x: np.array(x).astype(np.int))
-        gen_to_subid = extract_from_dict(dict_, "gen_to_subid", lambda x: np.array(x).astype(np.int))
-        lines_or_to_subid = extract_from_dict(dict_, "lines_or_to_subid", lambda x: np.array(x).astype(np.int))
-        lines_ex_to_subid = extract_from_dict(dict_, "lines_ex_to_subid", lambda x: np.array(x).astype(np.int))
-
-        load_to_sub_pos = extract_from_dict(dict_, "load_to_sub_pos", lambda x: np.array(x).astype(np.int))
-        gen_to_sub_pos = extract_from_dict(dict_, "gen_to_sub_pos", lambda x: np.array(x).astype(np.int))
-        lines_or_to_sub_pos = extract_from_dict(dict_, "lines_or_to_sub_pos", lambda x: np.array(x).astype(np.int))
-        lines_ex_to_sub_pos = extract_from_dict(dict_, "lines_ex_to_sub_pos", lambda x: np.array(x).astype(np.int))
-
-        load_pos_topo_vect = extract_from_dict(dict_, "load_pos_topo_vect", lambda x: np.array(x).astype(np.int))
-        gen_pos_topo_vect = extract_from_dict(dict_, "gen_pos_topo_vect", lambda x: np.array(x).astype(np.int))
-        lines_or_pos_topo_vect = extract_from_dict(dict_, "lines_or_pos_topo_vect", lambda x: np.array(x).astype(np.int))
-        lines_ex_pos_topo_vect = extract_from_dict(dict_, "lines_ex_pos_topo_vect", lambda x: np.array(x).astype(np.int))
-
-        observationClass_str = extract_from_dict(dict_, "observationClass", str)
-        observationClass_li = observationClass_str.split('.')
-
-        if observationClass_li[-1] in globals():
-            observationClass = globals()[observationClass_li[-1]]
-        else:
-            try:
-                observationClass = eval(observationClass_str)
-            except NameError:
-                msg_err_ = "Impossible to find the module \"{}\" to load back the observation space. Try \"from {} import {}\""
-                raise Grid2OpException(msg_err_.format(observationClass_str, ".".join(observationClass_li[:-1]), observationClass_li[-1]))
-            except AttributeError:
-                try:
-                    observationClass = eval(observationClass_li[-1])
-                except:
-                    if len(observationClass_li) > 1:
-                        msg_err_ = "Impossible to find the class named \"{}\" to load back the observation " \
-                                   "(module is found but not the class in it) Please import it via \"from {} import {}\"."
-                        msg_err_ = msg_err_.format(observationClass_str,
-                                                   ".".join(observationClass_li[:-1]),
-                                                   observationClass_li[-1])
-                    else:
-                        msg_err_ = "Impossible to import the class named \"{}\" to load back the observation space (the " \
-                                   "module is found but not the class in it)"
-                        msg_err_ = msg_err_.format(observationClass_str)
-                    raise Grid2OpException(msg_err_)
-
-        res = SerializableObservationSpace(n_gen, n_load, n_lines, subs_info,
-                                           load_to_subid, gen_to_subid, lines_or_to_subid, lines_ex_to_subid,
-                                           load_to_sub_pos, gen_to_sub_pos, lines_or_to_sub_pos, lines_ex_to_sub_pos,
-                                           load_pos_topo_vect, gen_pos_topo_vect, lines_or_pos_topo_vect,
-                                           lines_ex_pos_topo_vect,
-                                           observationClass=observationClass)
-        return res
-
-    def to_dict(self):
-        """
-        Serialize this object as a dictionnary.
-
-        Returns
-        -------
-        res: ``dict``
-            A dictionnary representing this object content. It can be loaded back with
-             :func:`SerializableObservationSpace.from_dict`
-        """
-        res = {}
-        save_to_dict(res, self, "n_gen", int)
-        save_to_dict(res, self, "n_load", int)
-        save_to_dict(res, self, "n_lines", int)
-        save_to_dict(res, self, "subs_info", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "load_to_subid", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "gen_to_subid", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "lines_or_to_subid", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "lines_ex_to_subid", lambda li: [int(el) for el in li])
-
-        save_to_dict(res, self, "load_to_sub_pos", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "gen_to_sub_pos", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "lines_or_to_sub_pos", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "lines_ex_to_sub_pos", lambda li: [int(el) for el in li])
-
-        save_to_dict(res, self, "load_pos_topo_vect", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "gen_pos_topo_vect", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "lines_or_pos_topo_vect", lambda li: [int(el) for el in li])
-        save_to_dict(res, self, "lines_ex_pos_topo_vect", lambda li: [int(el) for el in li])
-
-        save_to_dict(res, self, "observationClass", lambda x: re.sub("(<class ')|('>)", "", "{}".format(x)))
-
-        return res
-
-    def from_vect(self, obs):
-        """
-        Convert a observation, represented as a vector to a valid :class:`Observation` instance
-
-        Parameters
-        ----------
-        obs: ``numpy.ndarray``
-            The observation (represented as a numpy array) to convert to
-            :class:`grid2op.Observation.Observation` instance.
-
-        Returns
-        -------
-        res: :class:`grid2op.Observation.Observation`
-            The converted observation (converted from vector to :class:`grid2op.Observation.Observation` )
-
-        """
-        res = copy.deepcopy(self.empty_obs)
-        res.from_vect(obs)
-        return res
-
-    def get_obj_connect_to(self, _sentinel=None, substation_id=None):
-        """
-        Get all the object connected to a given substation:
-
-        Parameters
-        ----------
-        _sentinel: ``None``
-            Used to prevent positional parameters. Internal, do not use.
-
-        substation_id: ``int``
-            ID of the substation we want to inspect
-
-        Returns
-        -------
-        res: ``dict``
-            A dictionnary with keys:
-
-              - "loads_id": a vector giving the id of the loads connected to this substation, empty if none
-              - "generators_id": a vector giving the id of the generators connected to this substation, empty if none
-              - "lines_or_id": a vector giving the id of the origin end of the powerlines connected to this substation,
-                empty if none
-              - "lines_ex_id": a vector giving the id of the extermity end of the powerlines connected to this
-                substation, empty if none.
-              - "nb_elements" : number of elements connected to this substation
-
-        """
-
-        if substation_id is None:
-            raise Grid2OpException("You ask the composition of a substation without specifying its id."
-                                   "Please provide \"substation_id\"")
-        if substation_id >= len(self.subs_info):
-            raise Grid2OpException("There are no substation of id \"substation_id={}\" in this grid.".format(substation_id))
-
-        res = {}
-        res["loads_id"] = np.where(self.load_to_subid == substation_id)[0]
-        res["generators_id"] = np.where(self.gen_to_subid == substation_id)[0]
-        res["lines_or_id"] = np.where(self.lines_or_to_subid == substation_id)[0]
-        res["lines_ex_id"] = np.where(self.lines_ex_to_subid == substation_id)[0]
-        res["nb_elements"] = self.subs_info[substation_id]
-        return res
-
-    def get_lines_id(self, _sentinel=None, from_=None, to_=None):
-        """
-        Returns the list of all the powerlines id in the backend going from "from_" to "to_"
-
-        Parameters
-        ----------
-        _sentinel: ``None``
-            Internal, do not use
-
-        from_: ``int``
-            Id the substation to which the origin end of the powerline to look for should be connected to
-
-        to_: ``int``
-            Id the substation to which the extremity end of the powerline to look for should be connected to
-
-        Returns
-        -------
-        res: ``list``
-            Id of the powerline looked for.
-
-        Raises
-        ------
-        :class:`grid2op.Exceptions.BackendError` if no match is found.
-
-        """
-        res = []
-        if from_ is None:
-            raise BackendError("ObservationSpace.get_lines_id: impossible to look for a powerline with no origin substation. Please modify \"from_\" parameter")
-        if to_ is None:
-            raise BackendError("ObservationSpace.get_lines_id: impossible to look for a powerline with no extremity substation. Please modify \"to_\" parameter")
-
-        for i, (ori, ext) in enumerate(zip(self.lines_or_to_subid, self.lines_ex_to_subid)):
-            if ori == from_ and ext == to_:
-                res.append(i)
-
-        if res is []:
-            raise BackendError("ObservationSpace.get_line_id: impossible to find a powerline with connected at origin at {} and extremity at {}".format(from_, to_))
-
-        return res
-
-    def get_generators_id(self, sub_id):
-        """
-        Returns the list of all generators id in the backend connected to the substation sub_id
-
-        Parameters
-        ----------
-        sub_id: ``int``
-            The substation to which we look for the generator
-
-        Returns
-        -------
-        res: ``list``
-            Id of the generator id looked for.
-
-        Raises
-        ------
-        :class:`grid2op.Exceptions.BackendError` if no match is found.
-
-
-        """
-        res = []
-        if sub_id is None:
-            raise BackendError(
-                "ObservationSpace.get_generators_id: impossible to look for a generator not connected to any substation. Please modify \"sub_id\" parameter")
-
-        for i, s_id_gen in enumerate(self.gen_to_subid):
-            if s_id_gen == sub_id:
-                res.append(i)
-
-        if res is []:
-            raise BackendError(
-                "ObservationSpace.get_generators_id: impossible to find a generator connected at substation {}".format(sub_id))
-
-        return res
-
-    def get_loads_id(self, sub_id):
-        """
-        Returns the list of all generators id in the backend connected to the substation sub_id
-
-        Parameters
-        ----------
-        sub_id: ``int``
-            The substation to which we look for the generator
-
-        Returns
-        -------
-        res: ``list``
-            Id of the generator id looked for.
-
-        Raises
-        ------
-        :class:`grid2op.Exceptions.BackendError` if no match found.
-
-        """
-        res = []
-        if sub_id is None:
-            raise BackendError(
-                "ObservationSpace.get_loads_id: impossible to look for a load not connected to any substation. Please modify \"sub_id\" parameter")
-
-        for i, s_id_gen in enumerate(self.load_to_subid):
-            if s_id_gen == sub_id:
-                res.append(i)
-
-        if res is []:
-            raise BackendError(
-                "ObservationSpace.get_loads_id: impossible to find a load connected at substation {}".format(sub_id))
-
-        return res
-
-
-class ObservationHelper(SerializableObservationSpace):
+class ObservationHelper(SerializableSpace):
     """
     Helper that provides usefull functions to manipulate :class:`Observation`.
 
     Observation should only be built using this Helper. It is absolutely not recommended to make an observation
     directly form its constructor.
 
+    This class represents the same concept as the "Observation Space" in the OpenAI gym framework.
+
     Attributes
     ----------
+
+    observationClass: ``type``
+        Class used to build the observations. It defaults to :class:`CompleteObservation`
+
+    empty_obs: ``Observation.Observation``
+        An empty observation with the proper dimensions.
+
     parameters: :class:`grid2op.Parameters.Parameters`
         Type of Parameters used to compute powerflow for the forecast.
 
@@ -1886,7 +1492,7 @@ class ObservationHelper(SerializableObservationSpace):
 
     """
     def __init__(self,
-                 n_gen, n_load, n_lines, subs_info,
+                 name_prod, name_load, name_line, subs_info,
                  load_to_subid, gen_to_subid, lines_or_to_subid, lines_ex_to_subid,
                  load_to_sub_pos, gen_to_sub_pos, lines_or_to_sub_pos, lines_ex_to_sub_pos,
                  load_pos_topo_vect, gen_pos_topo_vect, lines_or_pos_topo_vect, lines_ex_pos_topo_vect,
@@ -1897,17 +1503,25 @@ class ObservationHelper(SerializableObservationSpace):
         Env: requires :attr:`grid2op.Environment.parameters` and :attr:`grid2op.Environment.backend` to be valid
         """
 
-        SerializableObservationSpace.__init__(self, n_gen, n_load, n_lines, subs_info,
-                 load_to_subid, gen_to_subid, lines_or_to_subid, lines_ex_to_subid,
-                 load_to_sub_pos, gen_to_sub_pos, lines_or_to_sub_pos, lines_ex_to_sub_pos,
-                 load_pos_topo_vect, gen_pos_topo_vect, lines_or_pos_topo_vect, lines_ex_pos_topo_vect,
-                                              observationClass=observationClass)
+        SerializableSpace.__init__(self,
+                                   name_prod=name_prod, name_load=name_load, name_line=name_line, subs_info=subs_info,
+                                   load_to_subid=load_to_subid, gen_to_subid=gen_to_subid,
+                                   lines_or_to_subid=lines_or_to_subid, lines_ex_to_subid=lines_ex_to_subid,
+                                   load_to_sub_pos=load_to_sub_pos, gen_to_sub_pos=gen_to_sub_pos,
+                                   lines_or_to_sub_pos=lines_or_to_sub_pos, lines_ex_to_sub_pos=lines_ex_to_sub_pos,
+                                   load_pos_topo_vect=load_pos_topo_vect, gen_pos_topo_vect=gen_pos_topo_vect,
+                                   lines_or_pos_topo_vect=lines_or_pos_topo_vect,
+                                   lines_ex_pos_topo_vect=lines_ex_pos_topo_vect,
+                                   subtype=observationClass)
+
+        self.observationClass = self.subtype
+        self.empty_obs = self.template_obj
 
         # TODO DOCUMENTATION !!!
 
         # print("ObservationHelper init with rewardClass: {}".format(rewardClass))
         self.parameters = copy.deepcopy(env.parameters)
-        # for the observation, I switch betwween the _parameters for the environment and for the simulation
+        # for the observation, I switch between the _parameters for the environment and for the simulation
         self.parameters.ENV_DC = self.parameters.FORECAST_DC
 
         if rewardClass is None:
@@ -1940,13 +1554,7 @@ class ObservationHelper(SerializableObservationSpace):
                                                obs_env=self.obs_env,
                                                action_helper=self.action_helper_env)
 
-        self.seed = None
-
     def __call__(self, env):
-        if self.seed is not None:
-            # in this case i have specific seed set. So i force the seed to be deterministic.
-            # TODO seed handling
-            self.seed = np.random.randint(4294967295)
         self.obs_env.update_grid(env.backend)
 
         res = self.observationClass(n_gen=self.n_gen, n_load=self.n_load, n_lines=self.n_lines,
@@ -1966,32 +1574,15 @@ class ObservationHelper(SerializableObservationSpace):
                                     seed=self.seed,
                                     obs_env=self.obs_env,
                                     action_helper=self.action_helper_env)
+
+        # TODO how to make sure that whatever the number of time i call "simulate" i still get the same observations
+        # TODO use self.obs_prng when updating actions
         res.update(env=env)
         return res
-
-    def seed(self, seed):
-        """
-        Use to set the seed in case of non determinitics observation.
-        :param seed:
-        :return:
-        """
-        self.seed = seed
 
     def size_obs(self):
         """
         Size if the observation vector would be flatten
         :return:
-        """
-        return self.n
-
-    def size(self):
-        """
-        Size if the observation vector would be flatten. That's also the dimension of the observation space.
-
-        Returns
-        -------
-        size: ``int``
-            The size defined above.
-
         """
         return self.n
